LPS-66054 Create new config contributors
